pyproject.toml
[project]
name = "photo-tv-sync"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "osxphotos>=0.68",
    "Pillow>=10.0",
    "pillow-heif>=0.18",
    "samsungtvws>=2.6",
]

[project.scripts]
photo-tv-sync = "photo_tv_sync.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/photo_tv_sync"]

